[libata promise] revert broken taskfile delivery change
The ->tf_load and ->exec_command driver hooks were changed to assume
that PIO was the only type of taskfile ever delivered to these functions.
This will be true... in the future, but not today.  In other drivers
this change was not needed, but Promise executes commands differently
due to its "ATA packet" hardware features, so the Promise drivers need
this change reverted.

Diagnosis and initial fix by Brad Campbell <brad@wasp.net.au>