[PATCH] vmscan: handle synchronous writepage()
Teach page reclaim to understand synchronous ->writepage implementations.

If ->writepage completed I/O prior to returning we can proceed to reclaim the
page without giving it another trip around the LRU.

This is beneficial for ramdisk-backed S_ISREG files: we can reclaim the file's
pages as fast as the ramdisk driver needs to allocate them and this prevents
I/O errors due to OOM in rd_blkdev_pagecache_IO().

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>