JFS: Don't allow reading beyond the inode map's EOF
If we try to read inodes that are beyond the size of the inode map,
__read_metapages would read unitialized pages into the inode map's
address space.  If the inode map is later grown in order to allocate
more inodes, the page is initialized and written under a different
address space.  Having the stale page in the page cache prevents the
properly initialized page from being read, and results in errors.

This problem can be provoked by an nfs client trying to read an inode
that does not exist.

Signed-off-by: Dave Kleikamp <shaggy@austin.ibm.com>