[PATCH] invalidate_inodes2(): mark pages not uptodate
Andrea Arcangeli <andrea@suse.de> points out that invalidate_inode_pages2() is
supposed to mark mapped-into-pagetable pages as not uptodate so that next time
someone faults the page in we will go get a new version from backing store.

The callers are the direct-io code and the NFS "something changed on the
server" code.  In both these cases we do need to go and re-read the page.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>