[PATCH] s390: lost dirty bits
The SetPageUptodate function is called for pages that are already up to
date.  The arch_set_page_uptodate function of s390 may not clear the dirty
bit in that case otherwise a dirty bit which is set between the start of an
i/o for a writeback and a following call to SetPageUptodate is lost.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>